requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
cartesia==2.0.17
certifi==2026.1.4
click==8.3.1
coloredlogs==15.0.1
daily-python==0.23.0
dataclasses-json==0.6.7
deepgram-sdk==4.7.0
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.128.0
flatbuffers==25.12.19
frozenlist==1.8.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
idna==3.11
iterators==0.2.0
jiter==0.12.0
joblib==1.5.3
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.10.1
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nltk==3.9.2
numba==0.61.2
numpy==2.2.6
onnxruntime==1.23.2
openai==2.16.0
packaging==26.0
pillow==11.3.0
pipecat-ai==0.0.100
propcache==0.4.1
protobuf==5.29.5
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyloudnorm==0.1.1
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
resampy==0.4.3
scipy==1.17.0
sniffio==1.3.1
soxr==0.5.0.post1
starlette==0.50.0
sympy==1.14.0
tqdm==4.67.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
